requirements.txt
Pillow == 7.0.0
openpyxl == 3.0.3
tqdm == 4.43.0
